pyproject.toml
[tool.poetry]
name = "telegram-cloudpayments-webhook"
version = "0.1.0"
description = ""
authors = ["BuildTools <unconfigured@null.spigotmc.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "2.25.1"
fastapi = "0.100.0"
uvicorn = "0.23.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
